# api/v2/openapi.yml
openapi: "3.0.2"
info:
  title: Authcore API
  version: "2.0"
paths:
...
  /api/v2/users/{id}/sessions:
    get:
      summary: Get user sessions.
      tags:
        - user
      parameters:
        - in: query
          name: limit
          description: Pagination Object limit
          schema:
            type: string
        - in: query
          name: page_token
          description: Pagination Token
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_size:
                    type: integer
                  next_page_token:
                    type: string
                  prev_page_token:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
...
components:
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: integer
        user_agent:
          type: string
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        last_seen_ip:
          type: string
        client_id:
          type: string
...